zjPyConAzureOps/requirements.txt
pymsteams
azure-mgmt-compute~=4.3